Add files via upload
This project is a comprehensive exploration of COVID-19 data focused on the United Arab Emirates (UAE). Utilizing SQL techniques like joins, Common Table Expressions (CTEs), window functions, and aggregate functions, it analyzes various aspects of the pandemic such as total cases, death rates, and vaccination coverage. Queries within the project aim to provide insights into the percentage of the population affected by COVID-19, the likelihood of death upon contracting the virus, and the progress of the vaccination campaign. Additionally, it features a global overview of new cases and deaths, and establishes a view for easy access to UAE vaccination data, taking into account the static nature of population data within the analysis period.